pyproject.toml
[project]
name = "speedtracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hurry-filesize>=0.9",
    "requests>=2.32.3",
]
